Add Program Manager view for XLS/PDF export & XLS import
