+++ assessments/templates/assessments/program_manager/learning_unit_score_encoding.html
@@ -156,15 +156,6 @@
                             <em class="fas fa-upload"></em> {% trans 'Inject XLS file' %}
                         </a>
                     </div>
-
-                    <a  id="lnk_submit"
-                        class="btn btn-warning pull-right trigger_modal"
-                        data-url="{{ learning_unit_submit_url }}"
-                        data-toggle="tooltip" data-placement="bottom"
-                        title="{% trans 'Definitive submission (Submit to faculty)' %}"
-                        role="button">
-                            <em class="fas fa-paper-plane"></em> {% trans 'Submission' %}
-                    </a>
                     {% include 'assessments/common/legend.html' %}
                 </div>
             </div>
