+++ assessments/views/common/score_sheet_pdf_export.py
@@ -23,15 +23,39 @@
 #    see http://www.gnu.org/licenses/.
 #
 ##############################################################################
-from assessments.views.common.score_sheet_pdf_export import ScoreSheetPDFExportBaseView
-from ddd.logic.encodage_des_notes.soumission.commands import GetFeuilleDeNotesCommand
+from django.utils.functional import cached_property
+from django.views import View
+
+from assessments.views.serializers.score_sheet import ScoreSheetPDFSerializer
+from ddd.logic.encodage_des_notes.soumission.commands import SearchAdressesFeuilleDeNotesCommand
 from infrastructure.messages_bus import message_bus_instance
+from osis_common.document import paper_sheet
+from osis_role.contrib.views import PermissionRequiredMixin
 
 
-class ScoreSheetPDFExportTutorView(ScoreSheetPDFExportBaseView):
+class ScoreSheetPDFExportBaseView(PermissionRequiredMixin, View):
+    # PermissionRequiredMixin
+    permission_required = "assessments.can_access_scoreencoding"
+
+    @cached_property
+    def person(self):
+        return self.request.user.person
+
+    def get(self, request, *args, **kwargs):
+        feuille_de_notes = self.get_feuille_de_notes()
+        donnees_administratives = self.get_donnees_administratives()
+
+        score_sheet_serialized = ScoreSheetPDFSerializer(instance={
+            'feuille_de_notes': feuille_de_notes,
+            'donnees_administratives': donnees_administratives
+        }, context={'person': self.person})
+        return paper_sheet.print_notes(score_sheet_serialized.data)
+
     def get_feuille_de_notes(self):
-        cmd = GetFeuilleDeNotesCommand(
-            matricule_fgs_enseignant=self.person.global_id,
-            code_unite_enseignement=self.kwargs['learning_unit_code'].upper()
+        raise NotImplementedError()
+
+    def get_donnees_administratives(self):
+        cmd = SearchAdressesFeuilleDeNotesCommand(
+            codes_unite_enseignement=[self.kwargs['learning_unit_code']]
         )
         return message_bus_instance.invoke(cmd)
