+++ assessments/views/program_manager/score_sheet_xls_export.py
@@ -23,15 +23,18 @@
 #    see http://www.gnu.org/licenses/.
 #
 ##############################################################################
-from assessments.views.common.score_sheet_pdf_export import ScoreSheetPDFExportBaseView
+from django.utils.functional import cached_property
+
+from assessments.views.common.score_sheet_xls_export import ScoreSheetXLSExportBaseView
 from ddd.logic.encodage_des_notes.soumission.commands import GetFeuilleDeNotesCommand
 from infrastructure.messages_bus import message_bus_instance
 
 
-class ScoreSheetPDFExportTutorView(ScoreSheetPDFExportBaseView):
-    def get_feuille_de_notes(self):
+class ScoreSheetXLSExportProgramManagerView(ScoreSheetXLSExportBaseView):
+    @cached_property
+    def feuille_de_notes(self):
         cmd = GetFeuilleDeNotesCommand(
             matricule_fgs_enseignant=self.person.global_id,
-            code_unite_enseignement=self.kwargs['learning_unit_code'].upper()
+            code_unite_enseignement=self.kwargs['learning_unit_code']
         )
         return message_bus_instance.invoke(cmd)
