+++ assessments/views/program_manager/score_sheet_xls_import.py
@@ -23,15 +23,9 @@
 #    see http://www.gnu.org/licenses/.
 #
 ##############################################################################
-from assessments.views.common.score_sheet_pdf_export import ScoreSheetPDFExportBaseView
-from ddd.logic.encodage_des_notes.soumission.commands import GetFeuilleDeNotesCommand
-from infrastructure.messages_bus import message_bus_instance
+from assessments.views.common.score_sheet_xls_import import ScoreSheetXLSImportBaseView
 
 
-class ScoreSheetPDFExportTutorView(ScoreSheetPDFExportBaseView):
-    def get_feuille_de_notes(self):
-        cmd = GetFeuilleDeNotesCommand(
-            matricule_fgs_enseignant=self.person.global_id,
-            code_unite_enseignement=self.kwargs['learning_unit_code'].upper()
-        )
-        return message_bus_instance.invoke(cmd)
+class ScoreSheetXLSImportProgramManagerView(ScoreSheetXLSImportBaseView):
+    def call_command(self, note_etudiant):
+        pass
