+++ assessments/views/score_encoding.py
@@ -51,8 +51,14 @@ from assessments.models import score_sheet_address as score_sheet_address_mdl
 from assessments.views.program_manager.learning_unit_score_encoding import LearningUnitScoreEncodingProgramManagerView
 from assessments.views.program_manager.learning_unit_score_encoding_form import \
     LearningUnitScoreEncodingProgramManagerFormView
+from assessments.views.program_manager.score_sheet_pdf_export import ScoreSheetPDFExportProgramManagerView
+from assessments.views.program_manager.score_sheet_xls_export import ScoreSheetXLSExportProgramManagerView
+from assessments.views.program_manager.score_sheet_xls_import import ScoreSheetXLSImportProgramManagerView
 from assessments.views.tutor.learning_unit_score_encoding import LearningUnitScoreEncodingTutorView
 from assessments.views.tutor.learning_unit_score_encoding_form import LearningUnitScoreEncodingTutorFormView
+from assessments.views.tutor.score_sheet_pdf_export import ScoreSheetPDFExportTutorView
+from assessments.views.tutor.score_sheet_xls_export import ScoreSheetXLSExportTutorView
+from assessments.views.tutor.score_sheet_xls_import import ScoreSheetXLSImportTutorView
 from attribution import models as mdl_attr
 from base import models as mdl
 from base.auth.roles import program_manager
@@ -824,4 +830,52 @@ class LearningUnitScoreEncodingFormView(LoginRequiredMixin, View):
             return LearningUnitScoreEncodingTutorFormView.as_view()(request, *args, **kwargs)
         elif EntityRoleHelper.has_role(self.person, ProgramManager):
             return LearningUnitScoreEncodingProgramManagerFormView.as_view()(request, *args, **kwargs)
+        return self.handle_no_permission()
+
+
+class ScoreSheetPDFExportView(LoginRequiredMixin, View):
+    @cached_property
+    def person(self):
+        return self.request.user.person
+
+    def dispatch(self, request, *args, **kwargs):
+        if not request.user.is_authenticated:
+            return self.handle_no_permission()
+
+        if EntityRoleHelper.has_role(self.person, Tutor):
+            return ScoreSheetPDFExportTutorView.as_view()(request, *args, **kwargs)
+        elif EntityRoleHelper.has_role(self.person, ProgramManager):
+            return ScoreSheetPDFExportProgramManagerView.as_view()(request, *args, **kwargs)
+        return self.handle_no_permission()
+
+
+class ScoreSheetXLSExportView(LoginRequiredMixin, View):
+    @cached_property
+    def person(self):
+        return self.request.user.person
+
+    def dispatch(self, request, *args, **kwargs):
+        if not request.user.is_authenticated:
+            return self.handle_no_permission()
+
+        if EntityRoleHelper.has_role(self.person, Tutor):
+            return ScoreSheetXLSExportTutorView.as_view()(request, *args, **kwargs)
+        elif EntityRoleHelper.has_role(self.person, ProgramManager):
+            return ScoreSheetXLSExportProgramManagerView.as_view()(request, *args, **kwargs)
+        return self.handle_no_permission()
+
+
+class ScoreSheetXLSImportView(LoginRequiredMixin, View):
+    @cached_property
+    def person(self):
+        return self.request.user.person
+
+    def dispatch(self, request, *args, **kwargs):
+        if not request.user.is_authenticated:
+            return self.handle_no_permission()
+
+        if EntityRoleHelper.has_role(self.person, Tutor):
+            return ScoreSheetXLSImportTutorView.as_view()(request, *args, **kwargs)
+        elif EntityRoleHelper.has_role(self.person, ProgramManager):
+            return ScoreSheetXLSImportProgramManagerView.as_view()(request, *args, **kwargs)
         return self.handle_no_permission()
